Fix turn off news and interests routine
If the "IsFeedsAvailable" and "EnableFeeds" registry entries are not
set, the "ShellFeedsTaskbarViewMode" entry will reset every time the
script is run by explorer.

There is an alternate method shared by u/Ill_Branch1293 on Reddit where
the explorer.eye is killed before the value is set, but I don't like
that approach because, well, it kills Windows Explorer and it might
interfere with running tasks. If you still want to see it, find it here:
https://www.reddit.com/r/sysadmin/comments/s1iomh/comment/hw0soe4/